Feature engineering
Next was feature engineering. I am trying to predict the tip amount. The transformation graph excel file contains a representation of my process. The model I made is about 50% accurate. I notice that it produces a less accurate model when I try to make one based on more recent data, so I will have to continue changing the way the model is created post semester. Various graphs in relation to my models accuracy are located in the folder labeled model in addition to the model itself.