data aggregation and visualizations
Lastly was data aggregation and visualization. Initially I wanted to aggregate every file together and then upload the finished aggregation to gcs, but the system actually ran out of ram repeatedly so I was forced to aggregate each file and upload it before moving on to the next. In retrospect I should have aggregated about five hundred thousand rows together at a time before uploading them to gcs. After that I proceed to make a series of graphs that attempt to draw insight into the data.